Separated ConclusionModal from Conclusion
