+++ apps/client/src/components/Conclusion.js
@@ -1,12 +1,4 @@
-import {
-  Button,
-  Checkbox,
-  Heading,
-  Label,
-  Paragraph,
-  Radio,
-  RadioGroup,
-} from "@datapunt/asc-ui";
+import { Button, Heading, Paragraph } from "@datapunt/asc-ui";
 import React, { Fragment } from "react";
 import { isIE, isMobile } from "react-device-detect";
 
@@ -18,13 +10,12 @@ import {
   PrintOnly,
 } from "../atoms";
 import { Olo } from "../config";
-import { topics } from "../config";
 import { actions, eventNames, sections } from "../config/matomo";
 import withTracking from "../hoc/withTracking";
 import { sttrOutcomes } from "../sttr_client/models/checker";
+import ConclusionModal from "./ConclusionModal";
 import ContactSentence from "./ContactSentence";
 import Markdown from "./Markdown";
-import Modal from "./Modal";
 
 const Conclusion = ({
   checker,
@@ -70,13 +61,6 @@ const Conclusion = ({
   const displayConclusions = contactConclusion
     ? [contactConclusion]
     : conclusions;
-
-  const handleModalButton = () => {
-    matomoTrackEvent({
-      action: actions.CLICK_INTERNAL_NAVIGATION,
-      name: eventNames.START_NEW_CHECK,
-    });
-  };
 
   const handlePermitButton = (e) => {
     e.preventDefault();
@@ -135,52 +119,7 @@ const Conclusion = ({
           </ComponentWrapper>
         )}
 
-        <Modal
-          buttonText="Nog een vergunningcheck doen"
-          heading="Kies een vergunningcheck"
-          onClick={handleModalButton}
-        >
-          <ComponentWrapper>
-            <Heading forwardedAs="h2">
-              Wilt u dezelfde adresgegevens gebruiken? [tekst updaten]
-            </Heading>
-            <Label htmlFor="ja" label="ja">
-              <Checkbox
-                value="value"
-                id="ja"
-                // onChange={(e) => onChange(e)}
-              />
-            </Label>
-          </ComponentWrapper>
-
-          <ComponentWrapper>
-            <Heading forwardedAs="h2">
-              Welke check wilt u doen? [tekst updaten]
-            </Heading>
-
-            <ComponentWrapper>
-              <RadioGroup name="checks">
-                {topics
-                  .filter((topic) => topic.sttrFile)
-                  .map((topic) => (
-                    <Label
-                      htmlFor={topic.name}
-                      label={topic.name}
-                      key={topic.name}
-                    >
-                      <Radio
-                        value={topic.slug}
-                        id={topic.name}
-                        // onChange={(e) => onChange(e)}
-                      />
-                    </Label>
-                  ))}
-              </RadioGroup>
-            </ComponentWrapper>
-
-            <Button variant="secondary">Gaaan [tekst updaten]</Button>
-          </ComponentWrapper>
-        </Modal>
+        <ConclusionModal />
       </HideForPrint>
 
       <PrintOnly withBorder avoidPageBreak>
